userMenu: Don't try to update an actor that may not be on the stage
The alternative-status-menu extension crashes the Shell because it removes this
actor from the menu, meaning that the "st_widget_get_theme_node" crashes the Shell.
One "more proper" general solution would be to create a variant of Signals.connect,
"connectActor", that makes sure that the signal doesn't run if its associated actor
is not on the stage.

https://bugzilla.gnome.org/show_bug.cgi?id=660520