shell-app: Add algorithm for grabbing icon color
This is a basic 3D histogram for pulling out the "most prominent color" from a
pixbuf. This will be used to color-code application-specific things based on
its icon.

Thanks a ton to pippin for giving me the details of the algorithm to work with,
explaining a lot of things and just being a nice person in general.

https://bugzilla.gnome.org/show_bug.cgi?id=661042