shell-global: Correct screenshot_window() after mutter changes
Recent mutter changes made MetaShapedTexture not a ClutterTexture,
but instead a special clutter actor that implemented the texture-y bits
itself. Use recently introduced API in MetaShapedTexture so that we can
get the raw texture data and spit it out as a PNG.

Use the new meta_shaped_texture_as_flattened_data() to get a window's
texture data, including any output shape or X SHAPE region it has,
allowing for a proper alpha channel on the screenshot.

https://bugzilla.gnome.org/show_bug.cgi?id=662486